+++ data/dev.py
@@ -5,16 +5,17 @@ from scenario import get_scenario_data
 from settings import DATA_PATH, DATA_SCENARIO_PATH
 
 
-def get_dummy_data(scenario_id):
-    with open(f"{DATA_PATH}/{DATA_SCENARIO_PATH}/{scenario_id}.json", 'r') as dummy_data_file:
+def get_dummy_data(scenario_id, table):
+    with open(f"{DATA_PATH}/{DATA_SCENARIO_PATH}/{scenario_id}_{table}.json", 'r') as dummy_data_file:
         data = json.load(dummy_data_file)
     return data
 
 
 def create_dummy_data(scenario_id):
-    data = get_scenario_data(scenario_id, as_text=True)
-    with open(f"{DATA_PATH}/{DATA_SCENARIO_PATH}/{scenario_id}.json", "w") as dummy_file:
-        dummy_file.write(data)
+    for table in ("oed_scalars", "oed_timeseries"):
+        data = get_scenario_data(scenario_id, table)
+        with open(f"{DATA_PATH}/{DATA_SCENARIO_PATH}/{scenario_id}_{table}.json", "w") as dummy_file:
+            json.dump(data, dummy_file)
 
 
 if __name__ == "__main__":